[refactor] Move `DebugModeManager` to single file
