+++ Features/Sources/Features/API/CheckInLoader.swift
@@ -80,41 +80,3 @@ extension DependencyValues {
     set { self[FirebaseCheckInLoader.self] = newValue }
   }
 }
-
-// MARK: - Debug Mode
-struct DebugModeManager {
-  var isFullAccess: (String) -> Bool
-  var setAccess: (String) -> Void
-}
-
-extension DebugModeManager: DependencyKey {
-  static var liveValue: DebugModeManager = DebugModeManager { key in
-    UserDefaults.standard.bool(forKey: key)
-  } setAccess: { activationKey in
-    let remoteConfig = RemoteConfig.remoteConfig()
-    remoteConfig.fetch { status, error in
-      if status == .success {
-        remoteConfig.activate { _, _ in
-          if let value = remoteConfig.configValue(forKey: "admin_full_access").stringValue, value == activationKey {
-            UserDefaults.standard.setValue(true, forKey: "admin_full_access")
-          }
-        }
-      }
-    }
-  }
-}
-
-extension DebugModeManager: TestDependencyKey {
-  static var testValue: DebugModeManager = DebugModeManager { _ in
-    unimplemented("DebugModeManager_isFullAccess")
-  } setAccess: { _ in
-    unimplemented("DebugModeManager_setAccess")
-  }
-}
-
-extension DependencyValues {
-  var debugModeManager: DebugModeManager {
-    get { self[DebugModeManager.self] }
-    set { self[DebugModeManager.self] = newValue }
-  }
-}
